[parsing] Extract ModelInstanceInfo to its own header file
Relevant to: #17390, #17112

This patch opens up the physical design of model directives, clearing
the way for integration of model directives as a language for Parser
(#17390), and porting of various modules from Anzu to Drake (#17112).

There should be no changes to either functionality or existing API
surface; only the arrangement of files and dependencies is changed.
