+++ multibody/parsing/process_model_directives.h
@@ -1,14 +1,13 @@
 #pragma once
 
-#include <optional>
 #include <string>
 #include <vector>
 
 #include "drake/multibody/parsing/model_directives.h"
+#include "drake/multibody/parsing/model_instance_info.h"
 #include "drake/multibody/parsing/package_map.h"
 #include "drake/multibody/parsing/parser.h"
 #include "drake/multibody/plant/multibody_plant.h"
-#include "drake/multibody/tree/multibody_tree_indexes.h"
 
 namespace drake {
 namespace multibody {
@@ -24,25 +23,6 @@ ModelDirectives LoadModelDirectives(const std::string& filename);
 std::string ResolveModelDirectiveUri(
     const std::string& uri,
     const drake::multibody::PackageMap& package_map);
-
-// TODO(#13074): Burn this in a dumpster fire pending real model
-// composition / extraction in Drake.
-// TODO(#14084): This structure might combine with the implementation of 14084:
-// https://github.com/RobotLocomotion/drake/issues/14084#issuecomment-694394869
-/// Convenience structure to hold all of the information to add a model
-/// instance from a file.
-struct ModelInstanceInfo {
-  /// Model name (possibly scoped).
-  std::string model_name;
-  /// File path.
-  std::string model_path;
-  /// WARNING: This is the *unscoped* parent frame, assumed to be unique.
-  std::string parent_frame_name;
-  /// This is the unscoped frame name belonging to `model_instance`.
-  std::string child_frame_name;
-  drake::math::RigidTransformd X_PC;
-  drake::multibody::ModelInstanceIndex model_instance;
-};
 
 /// Flatten model directives.
 void FlattenModelDirectives(const ModelDirectives& directives,
